refactor: add some hooks for encapsulate logic
