+++ src/components/CharacterItem/CharacterItem.module.scss
@@ -8,6 +8,12 @@
   border: 1px solid var(--color-gray-dark);
   color: var(--color-white-default);
   max-height: 500px;
+  cursor: pointer;
+  @extend %transition;
+
+  @include hover {
+    box-shadow: 1px 1px 7px var(--color-primary-bright);
+  }
 }
 
 .character-img-container {
