+++ src/hooks/useCard.ts
@@ -1,16 +1,15 @@
 import { CharacterWithFavorite, Film } from '@models/index';
-import { AppRoutes } from '@router/routes';
 import { useGetCharacterByIdQuery, useGetPlanetQuery } from '@store/api/swapiApi';
-import { fetchFavorites, toggleFavoriteInFirebase } from '@store/favoritesSlice';
-import { selectFavorites, selectFilms, selectUseIsLoggedIn } from '@store/selectors';
+import { fetchFavorites } from '@store/favoritesSlice';
+import { selectFavorites, selectFilms } from '@store/selectors';
 import { extractPlanetPath, isNotNullable, markFavorites } from '@utils/utils';
 import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { useAppDispatch } from './storeHooks';
+import { useHandleFavorites } from './useHandleFavorite';
 
 export const useCard = (callback: () => void) => {
-  const [isFavorite, setIsFavorite] = useState<boolean | null>(null);
   const dispatch = useAppDispatch();
   const favorites = useSelector(selectFavorites);
   const [preparedCharacter, setPreparedCharacter] = useState<CharacterWithFavorite | null>(null);
@@ -18,10 +17,11 @@ export const useCard = (callback: () => void) => {
   const [filteredFilms, setFilteredFilms] = useState<Film[]>([]);
   const [homeworld, setHomeworld] = useState<string | null>(null);
   const { id } = useParams();
-  const isLoggedIn = useSelector(selectUseIsLoggedIn);
-  const navigate = useNavigate();
+
   const { data: character } = useGetCharacterByIdQuery(+isNotNullable(id) || 0);
   const { data: planet } = useGetPlanetQuery(homeworld || '', { skip: !homeworld });
+
+  const { isLoggedIn, handleToggleFavorite } = useHandleFavorites(callback, id);
 
   useEffect(() => {
     const getFavorites = async () => {
@@ -35,27 +35,22 @@ export const useCard = (callback: () => void) => {
 
   useEffect(() => {
     const getData = async () => {
-      if (character && favorites && id && films) {
-        const preparedItem = markFavorites([character], favorites);
-        setPreparedCharacter(preparedItem[0]);
-        setIsFavorite(preparedItem[0].isFavorite);
-        setHomeworld(extractPlanetPath(preparedItem[0].homeworld));
+      if (character && id && films) {
+        setHomeworld(extractPlanetPath(character.homeworld));
         setFilteredFilms(films.filter(film => character.films.includes(film.url)));
+
+        if (isLoggedIn && favorites) {
+          const preparedItem = markFavorites([character], favorites);
+          setPreparedCharacter(preparedItem[0]);
+        } else if (!isLoggedIn) {
+          const preparedItem = markFavorites([character], []);
+          setPreparedCharacter(preparedItem[0]);
+        }
       }
     };
 
     getData();
-  }, [character, favorites, films, id]);
+  }, [character, favorites, films, id, isLoggedIn]);
 
-  const handleToggleFavorite = () => {
-    if (isLoggedIn && id) {
-      callback();
-      setIsFavorite(!isFavorite);
-      dispatch(toggleFavoriteInFirebase({ id }));
-    } else {
-      navigate(AppRoutes.LOGIN_ROUTE);
-    }
-  };
-
-  return { preparedCharacter, filteredFilms, planet, isLoggedIn, isFavorite, handleToggleFavorite };
+  return { preparedCharacter, filteredFilms, planet, isLoggedIn, handleToggleFavorite };
 };
